Configure the database using EFCore completely. Modify EntityRelationship.md.

+++ SqlContext/Interface/IBaseRecord.cs
@@ -2,7 +2,7 @@ namespace SqlContext.Interface;
 
 public interface IBaseRecord
 {
-    public ulong RowId { get; set; }
+    public int RowId { get; set; }
     public DateTime DateTime { get; set; }
     public double Money { get; set; }
     public string? Remark { get; set; }
